Inventory_Management_System/inventory_system/serializers/inventory_item.py
# serializers for inventory item model
from django.db import IntegrityError
from rest_framework import serializers
from ..models import InventoryItem


class InventoryItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = InventoryItem
        fields = "__all__"

    def create(self, validated_data):
        try:
            return super().create(validated_data)
        except IntegrityError:
            raise serializers.ValidationError("Item already exists")
